% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of the state to plot the crashes}

\item{year}{year of choice for plotting the accidents occurred}
}
\value{
this function returns the graph of the selected 
state from the year of interest.
}
\description{
this function takes the state number and plots 
the accidents per selected year.
}
\note{
as this function depends on make_filename 
and fars_read it is important to write the name 
of the file correctly and type the year properly, 
otherwise the function will return an error. 
In case there is no data for the selected state(integer) 
the function will stop and will throw an error, 
in case there are no accidents the function will 
notify through a message.
}
\examples{
fars_map_state(1,2013)

}
