% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Fars_read function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{of excel file whose extension is .csv.}
}
\value{
This function returns a data frame of class tbd_df, 
which is a special case of the base data.frame. This tibble 
format has the advantage of printing only the first 10 rows 
and columns of the file that fit on the screen.
}
\description{
This is a function that allows you to load all types of 
documents with .csv extension. You can load these documents
using the filename argument followed by the .csv extension 
inside double quotes and you will get a data frame in tibble 
format to work in the console.
}
\note{
You must make sure you have the file where the R 
directory is located or otherwise modify the directory, 
as this could generate an error; you must also write the 
file name correctly without omitting any characters and 
put the name between double quotes, otherwise it will 
generate an error and do not forget  to load the 
external packages.
}
\examples{
fars_read("directory/filename.csv")
fars_read("accident_2013.csv")
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
fars_read("accident_2015.csv.bz2")

}
