% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fars_summarize_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{Years}{a vector of numerical objects (years).}
}
\value{
this function returns a summarized data frame 
where the first column corresponds to the month and the 
following columns correspond to the count performed in 
the selected years.
}
\description{
function that takes a set of years and calls the 
function fars_read_years, then joins them together 
and at the end you get a summary of the data.
}
\note{
As this function depends on the function 
fars_read_years the error for the non-existence 
of the year will also be generated.
}
\examples{
fars_summarize_years(c(2013,2014))

}
