% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make_filename function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year found in the file name}
}
\value{
this function returns a formatted combination of 
input values, where the year is the input and the result 
is the string "accident_year.csv.bz2".
}
\description{
this function allows you to generate character objects 
containing a formatted combination of input values.
}
\note{
you must make sure to put a numeric object as input,
because if you put a text string you will get an error and 
a warning will come up if you put the text string between 
double quotes.
}
\examples{
make_filename(year)
make_filename(2013)
make_filename(2014)

}
